% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_level.R
\name{gen.level}
\alias{gen.level}
\title{Generate Levels}
\usage{
gen.level(x)
}
\arguments{
\item{x}{An integer greater than or equal to 2.}
}
\value{
a vector of integers as
         coded levels for a factor
         with total number of
         levels as  \code{x}.
}
\description{
Generate coded integers for given
total number of levels of a factor.
}
\examples{
# To generate 5 levels for a factor
gen.level(5)
}
