% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimal_factorial.R
\name{minimal.factorial}
\alias{minimal.factorial}
\title{Minimally Changed Run Sequences}
\usage{
minimal.factorial(z)
}
\arguments{
\item{z}{A vector of size 2 with entries integers greater than 1.}
}
\value{
returns minimally changed run sequences for given
factorial setting as \code{v}.
}
\description{
Generate minimally changed run sequences for a given asymmetrical
or symmetrical factorial design.
}
\examples{
#' ## make a vector of factor levels
z <- c(2,3,4)
# To generate minimmaly changed run sequence
minimal.factorial(z)
}
