% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minfact_2.R
\name{shwet}
\alias{shwet}
\title{Minimally Changed BBD}
\usage{
shwet(x, y)
}
\arguments{
\item{x}{A matrix of factor level combinations}

\item{y}{total number of levels of new facor}
}
\value{
returns a matrix of minimally changed run sequence
}
\description{
Generate Box Behnken design (BBD) with minimum level changes
in the run sequence.
}
\keyword{internal}

